+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.1.4
 
 - Diffrent note names added.
+
+## 0.2.0
+
+- Diffrent note names added.
+- Piano buttons callback added.